third_party/WebKit/LayoutTests/http/tests/images/image-currentsrc-broken.html

<!DOCTYPE html>
<script src="/js-test-resources/js-test.js"></script>
<body>
<script>
    document.body.innerHTML += '<img id="broken" srcset="http://127.0.0.1:8000/resources/slow-image.php?name=dummy.js&mimeType=image/png&sleep=100">';
    jsTestIsAsync = true;
    var img = window.document.getElementById("broken");
    img.onerror = function() {
        shouldBe('img.currentSrc', '"http://127.0.0.1:8000/resources/slow-image.php?name=dummy.js&mimeType=image/png&sleep=100"');
    };

    window.onload = function() {
        finishJSTest();
    };
</script>
</body>
