third_party/blink/web_tests/http/tests/images/image-currentsrc-broken-expected.txt

PASS img.currentSrc is "http://127.0.0.1:8000/resources/slow-image.php?name=dummy.js&mimeType=image/png&sleep=100"
PASS successfullyParsed is true

TEST COMPLETE
